requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7
redis==5.0.1

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==2.2.0
django-allauth==0.57.0

# API Documentation
drf-spectacular==0.26.5
drf-yasg==1.21.7

# Health Data Standards
fhirclient==3.2.0
hl7==0.4.5

# Data Processing & Analytics
pandas>=2.1.4
numpy>=1.26.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# Task Queue
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Caching
django-redis==5.4.0

# File Handling
Pillow==10.1.0
django-storages==1.14.2

# Environment & Configuration
python-decouple==3.8
django-environ==0.11.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Monitoring & Logging
sentry-sdk==1.38.0
django-debug-toolbar==4.2.0

# Internationalization
django-modeltranslation==0.18.12

# Health Specific
django-health-check==3.17.0
